Fix issue with insufficient data type
Covid is high. Number of people tested against it around the world
exceeded maximum value that Int is capable to hold.
Quick fix changes data type from Int to Long where necessary.

+++ app/src/main/java/com/mobilecoronatracker/model/GeneralReportModelable.kt
@@ -7,7 +7,7 @@ interface GeneralReportModelable {
     var deaths: Int
     var recovered: Int
     var critical: Int
-    var tests: Int
+    var tests: Long
     var todayCases: Int
     var todayDeaths: Int
     var casesPerMillion: Double
